Clouds now also get drawn outside the boundaries of the voxel world

+++ client/src/gfx/clouds.h
@@ -2,6 +2,6 @@
 #include "../../../common/src/common.h"
 
 void cloudsRender    ();
-void cloudsDraw      (u8 cx, u8 cy, u8 cz);
+void cloudsDraw      (int cx, int cy, int cz);
 void cloudsCalcColors();
 void cloudsInit      ();
